backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
twilio==9.2.3
# ML dependencies for head pose model
tensorflow>=2.15.0
numpy>=1.26.0
opencv-python>=4.8.0
Pillow>=10.0.0
# YOLO for multi-person detection
ultralytics>=8.2.0
# MediaPipe for head pose fallback (optional - has OpenCV fallback)
# mediapipe>=0.10.0
# Body visibility detection dependencies
scikit-learn>=1.3.0
